Pазделила код в main.js на модули
